Pull protected-main doctor flow out of the CLI entrypoint
The protected-main doctor path was still living inline inside src/cli/main.js, which kept sandbox lifecycle, merge-back, and output rendering coupled to the top-level command dispatcher. This commit moves that lifecycle into src/doctor/index.js, moves the remaining shared branch/config helpers into src/git/index.js, and keeps the doctor CLI behavior covered by focused regression tests.

Constraint: gx doctor protected-branch behavior and output had to stay stable during the extraction
Rejected: Fold the overlapping scaffold/DI extraction into this commit | unrelated branch drift was already present in the worktree
Confidence: medium
Scope-risk: moderate
Reversibility: clean
Directive: Do not move the protected-main doctor lifecycle or shared branch/config helpers back into src/cli/main.js; reconcile the overlapping scaffold/DI branch drift separately before finish
Tested: node --check src/cli/main.js src/doctor/index.js src/git/index.js; node --test test/cli-args-dispatch.test.js; node --test test/doctor.test.js; openspec validate agent-codex-extract-doctor-sandbox-module-2026-04-22-15-38 --type change --strict; openspec validate --specs
Not-tested: gx branch finish --branch agent/codex/extract-doctor-sandbox-module-2026-04-22-15-38 --base main --via-pr --wait-for-merge --cleanup; npm test
