+++ src/cli/main.js
@@ -4,6 +4,7 @@ const hooksModule = require('../hooks');
 const sandboxModule = require('../sandbox');
 const toolchainModule = require('../toolchain');
 const finishModule = require('../finish');
+const doctorModule = require('../doctor');
 const {
   fs,
   path,
@@ -72,6 +73,14 @@ const {
   resolveRepoRoot,
   isGitRepo,
   discoverNestedGitRepos,
+  gitRefExists,
+  readGitConfig,
+  currentBranchName,
+  ensureRepoBranch,
+  workingTreeIsDirty,
+  aheadBehind,
+  branchExists,
+  branchMergedIntoBase,
 } = require('../git');
 const {
   run,
@@ -1015,49 +1024,8 @@ function buildSandboxSetupArgs(options, sandboxTarget) {
   return args;
 }
 
-function buildSandboxDoctorArgs(options, sandboxTarget) {
-  const args = ['doctor', '--target', sandboxTarget];
-  if (options.dryRun) args.push('--dry-run');
-  appendForceArgs(args, options);
-  if (options.skipAgents) args.push('--skip-agents');
-  if (options.skipPackageJson) args.push('--skip-package-json');
-  if (options.skipGitignore) args.push('--no-gitignore');
-  if (!options.dropStaleLocks) args.push('--keep-stale-locks');
-  args.push(options.waitForMerge ? '--wait-for-merge' : '--no-wait-for-merge');
-  if (options.verboseAutoFinish) args.push('--verbose-auto-finish');
-  if (options.json) args.push('--json');
-  return args;
-}
-
 function isSpawnFailure(result) {
   return Boolean(result?.error) && typeof result?.status !== 'number';
-}
-
-function ensureRepoBranch(repoRoot, branch) {
-  const current = currentBranchName(repoRoot);
-  if (current === branch) {
-    return { ok: true, changed: false };
-  }
-
-  const checkoutResult = run('git', ['-C', repoRoot, 'checkout', branch], { timeout: 20_000 });
-  if (isSpawnFailure(checkoutResult)) {
-    return {
-      ok: false,
-      changed: false,
-      stdout: checkoutResult.stdout || '',
-      stderr: checkoutResult.stderr || '',
-    };
-  }
-  if (checkoutResult.status !== 0) {
-    return {
-      ok: false,
-      changed: false,
-      stdout: checkoutResult.stdout || '',
-      stderr: checkoutResult.stderr || '',
-    };
-  }
-
-  return { ok: true, changed: true };
 }
 
 function protectedBaseSandboxBranchPrefix() {
@@ -1076,10 +1044,6 @@ function protectedBaseSandboxBranchPrefix() {
 
 function protectedBaseSandboxWorktreePath(repoRoot, branchName) {
   return path.join(repoRoot, defaultAgentWorktreeRelativeDir(), branchName.replace(/\//g, '__'));
-}
-
-function gitRefExists(repoRoot, ref) {
-  return run('git', ['-C', repoRoot, 'show-ref', '--verify', '--quiet', ref]).status === 0;
 }
 
 function resolveProtectedBaseSandboxStartRef(repoRoot, baseBranch) {
@@ -1258,852 +1222,6 @@ function cleanupProtectedBaseSandbox(repoRoot, metadata) {
 
   result.note = 'sandbox worktree pruned';
   return result;
-}
-
-function parseGitPathList(output) {
-  return String(output || '')
-    .split('\n')
-    .map((line) => line.trim())
-    .filter((line) => line && line !== LOCK_FILE_RELATIVE);
-}
-
-function collectDoctorChangedPaths(worktreePath) {
-  const changed = new Set();
-  const commands = [
-    ['diff', '--name-only'],
-    ['diff', '--cached', '--name-only'],
-    ['ls-files', '--others', '--exclude-standard'],
-  ];
-  for (const gitArgs of commands) {
-    const result = run('git', ['-C', worktreePath, ...gitArgs], { timeout: 20_000 });
-    for (const filePath of parseGitPathList(result.stdout)) {
-      changed.add(filePath);
-    }
-  }
-  return Array.from(changed);
-}
-
-function collectDoctorDeletedPaths(worktreePath) {
-  const deleted = new Set();
-  const commands = [
-    ['diff', '--name-only', '--diff-filter=D'],
-    ['diff', '--cached', '--name-only', '--diff-filter=D'],
-  ];
-  for (const gitArgs of commands) {
-    const result = run('git', ['-C', worktreePath, ...gitArgs], { timeout: 20_000 });
-    for (const filePath of parseGitPathList(result.stdout)) {
-      deleted.add(filePath);
-    }
-  }
-  return Array.from(deleted);
-}
-
-function collectWorktreeDirtyPaths(worktreePath) {
-  const dirty = new Set();
-  const commands = [
-    ['diff', '--name-only'],
-    ['diff', '--cached', '--name-only'],
-    ['ls-files', '--others', '--exclude-standard'],
-  ];
-  for (const gitArgs of commands) {
-    const result = run('git', ['-C', worktreePath, ...gitArgs], { timeout: 20_000 });
-    for (const filePath of parseGitPathList(result.stdout)) {
-      dirty.add(filePath);
-    }
-  }
-  return Array.from(dirty);
-}
-
-function collectDoctorForceAddPaths(worktreePath) {
-  return REQUIRED_MANAGED_REPO_FILES
-    .filter((relativePath) => relativePath.startsWith('scripts/') || relativePath.startsWith('.githooks/'))
-    .filter((relativePath) => fs.existsSync(path.join(worktreePath, relativePath)));
-}
-
-function stripDoctorSandboxLocks(rawContent, branchName) {
-  if (!rawContent || !branchName) {
-    return rawContent;
-  }
-  try {
-    const parsed = JSON.parse(rawContent);
-    const locks = parsed && typeof parsed === 'object' && parsed.locks && typeof parsed.locks === 'object'
-      ? parsed.locks
-      : null;
-    if (!locks) {
-      return rawContent;
-    }
-    let changed = false;
-    const filteredLocks = {};
-    for (const [filePath, lockInfo] of Object.entries(locks)) {
-      if (lockInfo && lockInfo.branch === branchName) {
-        changed = true;
-        continue;
-      }
-      filteredLocks[filePath] = lockInfo;
-    }
-    if (!changed) {
-      return rawContent;
-    }
-    return `${JSON.stringify({ ...parsed, locks: filteredLocks }, null, 2)}\n`;
-  } catch {
-    return rawContent;
-  }
-}
-
-function claimDoctorChangedLocks(metadata) {
-  if (!metadata.branch) {
-    return {
-      status: 'skipped',
-      note: 'missing sandbox branch metadata',
-      changedCount: 0,
-      deletedCount: 0,
-    };
-  }
-
-  const changedPaths = Array.from(new Set([
-    ...collectDoctorChangedPaths(metadata.worktreePath),
-    ...collectDoctorForceAddPaths(metadata.worktreePath),
-  ]));
-  const deletedPaths = collectDoctorDeletedPaths(metadata.worktreePath);
-  if (changedPaths.length > 0) {
-    runPackageAsset('lockTool', ['claim', '--branch', metadata.branch, ...changedPaths], {
-      cwd: metadata.worktreePath,
-      timeout: 30_000,
-    });
-  }
-  if (deletedPaths.length > 0) {
-    runPackageAsset('lockTool', ['allow-delete', '--branch', metadata.branch, ...deletedPaths], {
-      cwd: metadata.worktreePath,
-      timeout: 30_000,
-    });
-  }
-
-  return {
-    status: 'claimed',
-    note: 'claimed locks for doctor auto-commit',
-    changedCount: changedPaths.length,
-    deletedCount: deletedPaths.length,
-  };
-}
-
-function autoCommitDoctorSandboxChanges(metadata) {
-  if (!metadata.worktreePath || !metadata.branch) {
-    return {
-      status: 'skipped',
-      note: 'missing sandbox branch metadata',
-    };
-  }
-
-  claimDoctorChangedLocks(metadata);
-  run(
-    'git',
-    ['-C', metadata.worktreePath, 'add', '-A', '--', '.', `:(exclude)${LOCK_FILE_RELATIVE}`],
-    { timeout: 20_000 },
-  );
-  const forceAddPaths = collectDoctorForceAddPaths(metadata.worktreePath);
-  if (forceAddPaths.length > 0) {
-    run(
-      'git',
-      ['-C', metadata.worktreePath, 'add', '-f', '--', ...forceAddPaths],
-      { timeout: 20_000 },
-    );
-  }
-  const staged = run(
-    'git',
-    ['-C', metadata.worktreePath, 'diff', '--cached', '--name-only', '--', '.', `:(exclude)${LOCK_FILE_RELATIVE}`],
-    { timeout: 20_000 },
-  );
-  const stagedFiles = parseGitPathList(staged.stdout);
-  if (stagedFiles.length === 0) {
-    return {
-      status: 'no-changes',
-      note: 'no committable doctor changes found in sandbox',
-    };
-  }
-
-  const commitResult = run(
-    'git',
-    ['-C', metadata.worktreePath, 'commit', '-m', 'Auto-finish: gx doctor repairs'],
-    { timeout: 30_000 },
-  );
-  if (commitResult.status !== 0) {
-    return {
-      status: 'failed',
-      note: 'doctor sandbox auto-commit failed',
-      stdout: commitResult.stdout || '',
-      stderr: commitResult.stderr || '',
-    };
-  }
-
-  return {
-    status: 'committed',
-    note: 'doctor sandbox repairs committed',
-    commitMessage: 'Auto-finish: gx doctor repairs',
-    stagedFiles,
-  };
-}
-
-function hasOriginRemote(repoRoot) {
-  return run('git', ['-C', repoRoot, 'remote', 'get-url', 'origin']).status === 0;
-}
-
-function originRemoteLooksLikeGithub(repoRoot) {
-  const originUrl = readGitConfig(repoRoot, 'remote.origin.url');
-  if (!originUrl) {
-    return false;
-  }
-  return /github\.com[:/]/i.test(originUrl);
-}
-
-function isCommandAvailable(commandName) {
-  return run('which', [commandName]).status === 0;
-}
-
-function extractAgentBranchFinishPrUrl(output) {
-  const match = String(output || '').match(/\[agent-branch-finish\] PR:\s*(\S+)/);
-  return match ? match[1] : '';
-}
-
-function doctorFinishFlowIsPending(output) {
-  return (
-    /\[agent-branch-finish\] PR merge not completed yet; leaving PR open\./.test(output) ||
-    /\[agent-branch-finish\] Merge pending review\/check policy\. Branch cleanup skipped for now\./.test(output) ||
-    /\[agent-branch-finish\] PR auto-merge enabled; waiting for required checks\/reviews\./.test(output)
-  );
-}
-
-function finishDoctorSandboxBranch(blocked, metadata, options = {}) {
-  if (!hasOriginRemote(blocked.repoRoot)) {
-    return {
-      status: 'skipped',
-      note: 'origin remote missing; skipped auto-finish',
-    };
-  }
-  const explicitGhBin = Boolean(String(process.env.GUARDEX_GH_BIN || '').trim());
-  if (!explicitGhBin && !originRemoteLooksLikeGithub(blocked.repoRoot)) {
-    return {
-      status: 'skipped',
-      note: 'origin remote is not GitHub; skipped auto-finish PR flow',
-    };
-  }
-
-  const ghBin = process.env.GUARDEX_GH_BIN || 'gh';
-  if (!isCommandAvailable(ghBin)) {
-    return {
-      status: 'skipped',
-      note: `'${ghBin}' not available; skipped auto-finish PR flow`,
-    };
-  }
-  const ghAuthStatus = run(ghBin, ['auth', 'status'], { timeout: 20_000 });
-  if (ghAuthStatus.status !== 0) {
-    return {
-      status: 'skipped',
-      note: `'${ghBin}' auth unavailable; skipped auto-finish PR flow`,
-      stderr: ghAuthStatus.stderr || '',
-    };
-  }
-
-  const rawWaitTimeoutSeconds = Number.parseInt(process.env.GUARDEX_FINISH_WAIT_TIMEOUT_SECONDS || '1800', 10);
-  const waitTimeoutSeconds =
-    Number.isFinite(rawWaitTimeoutSeconds) && rawWaitTimeoutSeconds >= 30 ? rawWaitTimeoutSeconds : 1800;
-  const finishTimeoutMs = Math.max(180_000, (waitTimeoutSeconds + 60) * 1000);
-  const waitForMergeArg = options.waitForMerge === false ? '--no-wait-for-merge' : '--wait-for-merge';
-
-  const finishResult = runPackageAsset(
-    'branchFinish',
-    ['--branch', metadata.branch, '--base', blocked.branch, '--via-pr', waitForMergeArg, '--cleanup'],
-    { cwd: metadata.worktreePath, timeout: finishTimeoutMs },
-  );
-  if (isSpawnFailure(finishResult)) {
-    return {
-      status: 'failed',
-      note: 'doctor sandbox finish flow errored',
-      stdout: finishResult.stdout || '',
-      stderr: finishResult.stderr || '',
-    };
-  }
-  if (finishResult.status !== 0) {
-    return {
-      status: 'failed',
-      note: 'doctor sandbox finish flow failed',
-      stdout: finishResult.stdout || '',
-      stderr: finishResult.stderr || '',
-    };
-  }
-
-  const combinedOutput = `${finishResult.stdout || ''}\n${finishResult.stderr || ''}`;
-  if (doctorFinishFlowIsPending(combinedOutput)) {
-    return {
-      status: 'pending',
-      note: 'PR created and waiting for merge policy/checks',
-      prUrl: extractAgentBranchFinishPrUrl(combinedOutput),
-      stdout: finishResult.stdout || '',
-      stderr: finishResult.stderr || '',
-    };
-  }
-
-  return {
-    status: 'completed',
-    note: 'doctor sandbox finish flow completed',
-    stdout: finishResult.stdout || '',
-    stderr: finishResult.stderr || '',
-  };
-}
-
-function mergeDoctorSandboxRepairsBackToProtectedBase(options, blocked, metadata, autoCommitResult, finishResult) {
-  if (options.dryRun) {
-    return {
-      status: autoCommitResult.status === 'committed' ? 'would-merge' : 'skipped',
-      note: autoCommitResult.status === 'committed'
-        ? 'dry run: would fast-forward tracked doctor repairs into the protected base workspace'
-        : 'dry run skips tracked repair merge',
-    };
-  }
-
-  if (autoCommitResult.status !== 'committed') {
-    return {
-      status: autoCommitResult.status === 'no-changes' ? 'unchanged' : 'skipped',
-      note: autoCommitResult.status === 'no-changes'
-        ? 'no tracked doctor repairs needed in the protected base workspace'
-        : 'tracked doctor repair merge skipped',
-    };
-  }
-
-  if (finishResult.status !== 'skipped') {
-    return {
-      status: 'skipped',
-      note: finishResult.status === 'failed'
-        ? 'tracked doctor repairs remain in the sandbox after finish failure'
-        : 'tracked doctor repairs are being delivered through the sandbox finish flow',
-    };
-  }
-
-  const allowedPaths = new Set([
-    ...(autoCommitResult.stagedFiles || []),
-    ...OMX_SCAFFOLD_DIRECTORIES,
-    ...Array.from(OMX_SCAFFOLD_FILES.keys()),
-    ...REQUIRED_MANAGED_REPO_FILES,
-    'bin',
-    'package.json',
-    '.gitignore',
-    'AGENTS.md',
-  ]);
-  const dirtyPaths = collectWorktreeDirtyPaths(blocked.repoRoot);
-  let stashRef = '';
-  if (dirtyPaths.length > 0) {
-    const unexpectedPaths = dirtyPaths.filter((filePath) => {
-      if (allowedPaths.has(filePath)) {
-        return false;
-      }
-      return !AGENT_WORKTREE_RELATIVE_DIRS.some(
-        (relativeDir) => filePath === relativeDir || filePath.startsWith(`${relativeDir}/`),
-      );
-    });
-    if (unexpectedPaths.length > 0) {
-      return {
-        status: 'failed',
-        note: `protected branch workspace has unrelated local changes: ${unexpectedPaths.join(', ')}`,
-      };
-    }
-    const stashMessage = `guardex-doctor-merge-${Date.now()}`;
-    const stashResult = run(
-      'git',
-      ['-C', blocked.repoRoot, 'stash', 'push', '--all', '--message', stashMessage],
-      { timeout: 30_000 },
-    );
-    if (isSpawnFailure(stashResult)) {
-      return {
-        status: 'failed',
-        note: 'could not stash protected branch doctor drift before merge',
-        stdout: stashResult.stdout || '',
-        stderr: stashResult.stderr || '',
-      };
-    }
-    if (stashResult.status !== 0) {
-      return {
-        status: 'failed',
-        note: 'stashing protected branch doctor drift failed',
-        stdout: stashResult.stdout || '',
-        stderr: stashResult.stderr || '',
-      };
-    }
-
-    const stashLookup = run(
-      'git',
-      ['-C', blocked.repoRoot, 'stash', 'list'],
-      { timeout: 20_000 },
-    );
-    stashRef = String(stashLookup.stdout || '')
-      .split('\n')
-      .find((line) => line.includes(stashMessage))
-      ?.split(':')[0]
-      ?.trim() || '';
-  }
-
-  const restoreResult = ensureRepoBranch(blocked.repoRoot, blocked.branch);
-  if (!restoreResult.ok) {
-    if (stashRef) {
-      run('git', ['-C', blocked.repoRoot, 'stash', 'apply', stashRef], { timeout: 30_000 });
-    }
-    return {
-      status: 'failed',
-      note: `could not restore protected branch '${blocked.branch}' before applying sandbox repairs`,
-      stdout: restoreResult.stdout || '',
-      stderr: restoreResult.stderr || '',
-    };
-  }
-
-  const mergeResult = run(
-    'git',
-    ['-C', blocked.repoRoot, 'merge', '--ff-only', metadata.branch],
-    { timeout: 30_000 },
-  );
-  if (isSpawnFailure(mergeResult)) {
-    if (stashRef) {
-      run('git', ['-C', blocked.repoRoot, 'stash', 'apply', stashRef], { timeout: 30_000 });
-    }
-    return {
-      status: 'failed',
-      note: 'tracked doctor repair merge errored',
-      stdout: mergeResult.stdout || '',
-      stderr: mergeResult.stderr || '',
-    };
-  }
-  if (mergeResult.status !== 0) {
-    if (stashRef) {
-      run('git', ['-C', blocked.repoRoot, 'stash', 'apply', stashRef], { timeout: 30_000 });
-    }
-    return {
-      status: 'failed',
-      note: 'tracked doctor repair merge failed',
-      stdout: mergeResult.stdout || '',
-      stderr: mergeResult.stderr || '',
-    };
-  }
-
-  let cleanupResult;
-  try {
-    cleanupResult = cleanupProtectedBaseSandbox(blocked.repoRoot, metadata);
-  } catch (error) {
-    return {
-      status: 'failed',
-      note: `tracked doctor repair merge succeeded but sandbox cleanup failed: ${error.message}`,
-      stdout: mergeResult.stdout || '',
-      stderr: mergeResult.stderr || '',
-    };
-  }
-
-  let hookRefreshResult;
-  try {
-    hookRefreshResult = configureHooks(blocked.repoRoot, false);
-  } catch (error) {
-    return {
-      status: 'failed',
-      note: `tracked doctor repair merge succeeded but local hook refresh failed: ${error.message}`,
-      stdout: mergeResult.stdout || '',
-      stderr: mergeResult.stderr || '',
-    };
-  }
-
-  if (stashRef) {
-    run('git', ['-C', blocked.repoRoot, 'stash', 'drop', stashRef], { timeout: 20_000 });
-  }
-
-  return {
-    status: 'merged',
-    note: 'fast-forwarded tracked doctor repairs into the protected base workspace',
-    stdout: mergeResult.stdout || '',
-    stderr: mergeResult.stderr || '',
-    cleanup: cleanupResult,
-    hookRefresh: hookRefreshResult,
-  };
-}
-
-/**
- * @param {string} [note]
- * @returns {OperationResult}
- */
-function createDoctorSkippedOperation(note = 'sandbox doctor did not complete successfully') {
-  return {
-    status: 'skipped',
-    note,
-  };
-}
-
-/**
- * @param {string} [note]
- * @returns {AutoFinishSummary}
- */
-function createSkippedDoctorAutoFinishSummary(note = 'sandbox doctor did not complete successfully') {
-  return {
-    enabled: false,
-    attempted: 0,
-    completed: 0,
-    skipped: 0,
-    failed: 0,
-    details: [`Skipped auto-finish sweep (${note}).`],
-  };
-}
-
-/**
- * Default the lifecycle to skipped states until the nested doctor run succeeds.
- *
- * @param {string} [note]
- * @returns {DoctorSandboxExecution}
- */
-function createDoctorSandboxExecutionState(note = 'sandbox doctor did not complete successfully') {
-  return {
-    autoCommit: createDoctorSkippedOperation(note),
-    finish: createDoctorSkippedOperation(note),
-    protectedBaseRepairSync: createDoctorSkippedOperation(note),
-    lockSync: createDoctorSkippedOperation(note),
-    omxScaffoldSync: createDoctorSkippedOperation(note),
-    autoFinish: createSkippedDoctorAutoFinishSummary(note),
-    sandboxLockContent: null,
-  };
-}
-
-/**
- * @param {string} repoRoot
- * @param {boolean} dryRun
- * @returns {OperationResult}
- */
-function summarizeDoctorOmxScaffoldSync(repoRoot, dryRun) {
-  const omxScaffoldOps = ensureOmxScaffold(repoRoot, dryRun);
-  const changedOmxPaths = omxScaffoldOps.filter((operation) => operation.status !== 'unchanged');
-  if (changedOmxPaths.length === 0) {
-    return {
-      status: 'unchanged',
-      note: '.omx scaffold already in sync',
-      operations: omxScaffoldOps,
-    };
-  }
-  return {
-    status: dryRun ? 'would-sync' : 'synced',
-    note: `${dryRun ? 'would sync' : 'synced'} ${changedOmxPaths.length} .omx path(s)`,
-    operations: omxScaffoldOps,
-  };
-}
-
-/**
- * @param {string} repoRoot
- * @param {SandboxMetadata} metadata
- * @returns {DoctorLockSyncState}
- */
-function syncDoctorLockRegistryBeforeMerge(repoRoot, metadata) {
-  const sandboxLockPath = path.join(metadata.worktreePath, LOCK_FILE_RELATIVE);
-  const baseLockPath = path.join(repoRoot, LOCK_FILE_RELATIVE);
-  if (!fs.existsSync(baseLockPath)) {
-    return {
-      result: {
-        status: 'skipped',
-        note: `${LOCK_FILE_RELATIVE} missing in protected base workspace`,
-      },
-      sandboxLockContent: null,
-    };
-  }
-  if (!fs.existsSync(sandboxLockPath)) {
-    return {
-      result: {
-        status: 'skipped',
-        note: `${LOCK_FILE_RELATIVE} missing in sandbox worktree`,
-      },
-      sandboxLockContent: null,
-    };
-  }
-
-  const sourceContent = stripDoctorSandboxLocks(
-    fs.readFileSync(sandboxLockPath, 'utf8'),
-    metadata.branch,
-  );
-  const destinationContent = fs.readFileSync(baseLockPath, 'utf8');
-  if (sourceContent === destinationContent) {
-    return {
-      result: {
-        status: 'unchanged',
-        note: `${LOCK_FILE_RELATIVE} already in sync`,
-      },
-      sandboxLockContent: sourceContent,
-    };
-  }
-
-  fs.mkdirSync(path.dirname(baseLockPath), { recursive: true });
-  fs.writeFileSync(baseLockPath, sourceContent, 'utf8');
-  return {
-    result: {
-      status: 'synced',
-      note: `${LOCK_FILE_RELATIVE} synced from sandbox`,
-    },
-    sandboxLockContent: sourceContent,
-  };
-}
-
-/**
- * @param {string} repoRoot
- * @param {string | null} sandboxLockContent
- * @returns {OperationResult}
- */
-function syncDoctorLockRegistryAfterMerge(repoRoot, sandboxLockContent) {
-  if (sandboxLockContent === null) {
-    return {
-      status: 'skipped',
-      note: `${LOCK_FILE_RELATIVE} missing in sandbox worktree`,
-    };
-  }
-
-  const baseLockPath = path.join(repoRoot, LOCK_FILE_RELATIVE);
-  if (!fs.existsSync(baseLockPath)) {
-    fs.mkdirSync(path.dirname(baseLockPath), { recursive: true });
-    fs.writeFileSync(baseLockPath, sandboxLockContent, 'utf8');
-    return {
-      status: 'synced',
-      note: `${LOCK_FILE_RELATIVE} recreated from sandbox`,
-    };
-  }
-
-  const destinationContent = fs.readFileSync(baseLockPath, 'utf8');
-  if (sandboxLockContent === destinationContent) {
-    return {
-      status: 'unchanged',
-      note: `${LOCK_FILE_RELATIVE} already in sync`,
-    };
-  }
-
-  fs.mkdirSync(path.dirname(baseLockPath), { recursive: true });
-  fs.writeFileSync(baseLockPath, sandboxLockContent, 'utf8');
-  return {
-    status: 'synced',
-    note: `${LOCK_FILE_RELATIVE} synced from sandbox`,
-  };
-}
-
-/**
- * @param {object} options
- * @param {{ repoRoot: string, branch: string }} blocked
- * @param {SandboxMetadata} metadata
- * @returns {DoctorSandboxExecution}
- */
-function executeDoctorSandboxLifecycle(options, blocked, metadata) {
-  const execution = createDoctorSandboxExecutionState();
-  const dryRun = Boolean(options.dryRun);
-
-  execution.omxScaffoldSync = summarizeDoctorOmxScaffoldSync(blocked.repoRoot, dryRun);
-
-  if (!dryRun) {
-    execution.autoCommit = autoCommitDoctorSandboxChanges(metadata);
-    if (execution.autoCommit.status === 'committed') {
-      execution.finish = finishDoctorSandboxBranch(blocked, metadata, options);
-    } else if (execution.autoCommit.status === 'no-changes') {
-      execution.finish = createDoctorSkippedOperation('no doctor changes to auto-finish');
-    } else if (execution.autoCommit.status !== 'failed') {
-      execution.finish = createDoctorSkippedOperation('auto-commit did not run');
-    }
-  } else {
-    execution.autoCommit = createDoctorSkippedOperation('dry-run skips doctor sandbox auto-commit');
-    execution.finish = createDoctorSkippedOperation('dry-run skips doctor sandbox finish flow');
-  }
-
-  const lockSyncState = syncDoctorLockRegistryBeforeMerge(blocked.repoRoot, metadata);
-  execution.lockSync = lockSyncState.result;
-  execution.sandboxLockContent = lockSyncState.sandboxLockContent;
-
-  execution.protectedBaseRepairSync = mergeDoctorSandboxRepairsBackToProtectedBase(
-    options,
-    blocked,
-    metadata,
-    execution.autoCommit,
-    execution.finish,
-  );
-
-  execution.omxScaffoldSync = summarizeDoctorOmxScaffoldSync(blocked.repoRoot, dryRun);
-  execution.lockSync = syncDoctorLockRegistryAfterMerge(
-    blocked.repoRoot,
-    execution.sandboxLockContent,
-  );
-  execution.autoFinish = autoFinishReadyAgentBranches(blocked.repoRoot, {
-    baseBranch: blocked.branch,
-    dryRun: options.dryRun,
-    waitForMerge: options.waitForMerge,
-    excludeBranches: [metadata.branch],
-  });
-
-  return execution;
-}
-
-function emitDoctorSandboxJsonOutput(nestedResult, execution) {
-  if (nestedResult.stdout) {
-    if (nestedResult.status === 0) {
-      try {
-        const parsed = JSON.parse(nestedResult.stdout);
-        process.stdout.write(
-          JSON.stringify(
-            {
-              ...parsed,
-              protectedBaseRepairSync: execution.protectedBaseRepairSync,
-              sandboxOmxScaffoldSync: execution.omxScaffoldSync,
-              sandboxLockSync: execution.lockSync,
-              sandboxAutoCommit: execution.autoCommit,
-              sandboxFinish: execution.finish,
-              autoFinish: execution.autoFinish,
-            },
-            null,
-            2,
-          ) + '\n',
-        );
-      } catch {
-        process.stdout.write(nestedResult.stdout);
-      }
-    } else {
-      process.stdout.write(nestedResult.stdout);
-    }
-  }
-  if (nestedResult.stderr) process.stderr.write(nestedResult.stderr);
-}
-
-/**
- * @param {object} options
- * @param {{ branch: string }} blocked
- * @param {SandboxMetadata} metadata
- * @param {SandboxStartResult} startResult
- * @param {any} nestedResult
- * @param {DoctorSandboxExecution} execution
- */
-function emitDoctorSandboxConsoleOutput(options, blocked, metadata, startResult, nestedResult, execution) {
-  console.log(
-    `[${TOOL_NAME}] doctor detected protected branch '${blocked.branch}'. ` +
-    `Running repairs in sandbox branch '${metadata.branch || 'agent/<auto>'}'.`,
-  );
-  if (startResult.stdout) process.stdout.write(startResult.stdout);
-  if (startResult.stderr) process.stderr.write(startResult.stderr);
-  if (nestedResult.stdout) process.stdout.write(nestedResult.stdout);
-  if (nestedResult.stderr) process.stderr.write(nestedResult.stderr);
-  if (nestedResult.status !== 0) {
-    return;
-  }
-
-  if (execution.autoCommit.status === 'committed') {
-    console.log(
-      `[${TOOL_NAME}] Auto-committed doctor repairs in sandbox branch '${metadata.branch}'.`,
-    );
-  } else if (execution.autoCommit.status === 'failed') {
-    console.log(`[${TOOL_NAME}] Doctor sandbox auto-commit failed; branch left for manual follow-up.`);
-    if (execution.autoCommit.stdout) process.stdout.write(execution.autoCommit.stdout);
-    if (execution.autoCommit.stderr) process.stderr.write(execution.autoCommit.stderr);
-  } else {
-    console.log(`[${TOOL_NAME}] Doctor sandbox auto-commit skipped: ${execution.autoCommit.note}.`);
-  }
-
-  if (execution.protectedBaseRepairSync.status === 'merged') {
-    console.log(`[${TOOL_NAME}] Fast-forwarded tracked doctor repairs into the protected branch workspace.`);
-  } else if (execution.protectedBaseRepairSync.status === 'unchanged') {
-    console.log(`[${TOOL_NAME}] Protected branch workspace already had the tracked doctor repairs.`);
-  } else if (execution.protectedBaseRepairSync.status === 'would-merge') {
-    console.log(`[${TOOL_NAME}] Dry run: would fast-forward tracked doctor repairs into the protected branch workspace.`);
-  } else if (execution.protectedBaseRepairSync.status === 'failed') {
-    console.log(`[${TOOL_NAME}] Protected branch tracked repair merge failed: ${execution.protectedBaseRepairSync.note}.`);
-    if (execution.protectedBaseRepairSync.stdout) process.stdout.write(execution.protectedBaseRepairSync.stdout);
-    if (execution.protectedBaseRepairSync.stderr) process.stderr.write(execution.protectedBaseRepairSync.stderr);
-  } else {
-    console.log(`[${TOOL_NAME}] Protected branch tracked repair merge skipped: ${execution.protectedBaseRepairSync.note}.`);
-  }
-
-  if (execution.lockSync.status === 'synced') {
-    console.log(
-      `[${TOOL_NAME}] Synced repaired lock registry back to protected branch workspace (${LOCK_FILE_RELATIVE}).`,
-    );
-  } else if (execution.lockSync.status === 'unchanged') {
-    console.log(`[${TOOL_NAME}] Lock registry already synced in protected branch workspace.`);
-  } else {
-    console.log(`[${TOOL_NAME}] Lock registry sync skipped: ${execution.lockSync.note}.`);
-  }
-
-  if (execution.finish.status === 'completed') {
-    console.log(`[${TOOL_NAME}] Auto-finish flow completed for sandbox branch '${metadata.branch}'.`);
-    if (execution.finish.stdout) process.stdout.write(execution.finish.stdout);
-    if (execution.finish.stderr) process.stderr.write(execution.finish.stderr);
-  } else if (execution.finish.status === 'pending') {
-    console.log(
-      `[${TOOL_NAME}] Auto-finish pending for sandbox branch '${metadata.branch}': ${execution.finish.note}.`,
-    );
-    if (execution.finish.prUrl) {
-      console.log(`[${TOOL_NAME}] PR: ${execution.finish.prUrl}`);
-    }
-    if (execution.finish.stdout) process.stdout.write(execution.finish.stdout);
-    if (execution.finish.stderr) process.stderr.write(execution.finish.stderr);
-  } else if (execution.finish.status === 'failed') {
-    console.log(`[${TOOL_NAME}] Auto-finish flow failed for sandbox branch '${metadata.branch}'.`);
-    if (execution.finish.stdout) process.stdout.write(execution.finish.stdout);
-    if (execution.finish.stderr) process.stderr.write(execution.finish.stderr);
-  } else {
-    console.log(`[${TOOL_NAME}] Auto-finish skipped: ${execution.finish.note}.`);
-  }
-
-  printAutoFinishSummary(execution.autoFinish, {
-    baseBranch: blocked.branch,
-    verbose: options.verboseAutoFinish,
-  });
-  if (execution.omxScaffoldSync.status === 'synced') {
-    console.log(`[${TOOL_NAME}] Synced .omx scaffold back to protected branch workspace.`);
-  } else if (execution.omxScaffoldSync.status === 'unchanged') {
-    console.log(`[${TOOL_NAME}] .omx scaffold already aligned in protected branch workspace.`);
-  } else if (execution.omxScaffoldSync.status === 'would-sync') {
-    console.log(`[${TOOL_NAME}] Dry run: would sync .omx scaffold back to protected branch workspace.`);
-  } else {
-    console.log(`[${TOOL_NAME}] .omx scaffold sync skipped: ${execution.omxScaffoldSync.note}.`);
-  }
-}
-
-function setDoctorSandboxExitCode(nestedResult, execution) {
-  if (typeof nestedResult.status === 'number') {
-    let exitCode = nestedResult.status;
-    if (exitCode === 0 && execution.autoCommit.status === 'failed') {
-      exitCode = 1;
-    }
-    if (
-      exitCode === 0 &&
-      execution.autoCommit.status === 'committed' &&
-      (execution.finish.status === 'failed' || execution.finish.status === 'pending')
-    ) {
-      exitCode = 1;
-    }
-    if (exitCode === 0 && execution.protectedBaseRepairSync.status === 'failed') {
-      exitCode = 1;
-    }
-    process.exitCode = exitCode;
-    return;
-  }
-  process.exitCode = 1;
-}
-
-function runDoctorInSandbox(options, blocked) {
-  /** @type {SandboxStartResult} */
-  const startResult = startProtectedBaseSandbox(blocked, {
-    taskName: `${SHORT_TOOL_NAME}-doctor`,
-    sandboxSuffix: 'gx-doctor',
-  });
-  const metadata = startResult.metadata;
-
-  const sandboxTarget = resolveSandboxTarget(blocked.repoRoot, metadata.worktreePath, options.target);
-  const nestedResult = run(
-    process.execPath,
-    [__filename, ...buildSandboxDoctorArgs(options, sandboxTarget)],
-    { cwd: metadata.worktreePath },
-  );
-  if (isSpawnFailure(nestedResult)) {
-    throw nestedResult.error;
-  }
-
-  const execution = nestedResult.status === 0
-    ? executeDoctorSandboxLifecycle(options, blocked, metadata)
-    : createDoctorSandboxExecutionState();
-
-  if (options.json) {
-    emitDoctorSandboxJsonOutput(nestedResult, execution);
-  } else {
-    emitDoctorSandboxConsoleOutput(options, blocked, metadata, startResult, nestedResult, execution);
-  }
-
-  setDoctorSandboxExitCode(nestedResult, execution);
 }
 
 function runSetupInSandbox(options, blocked, repoLabel = '') {
@@ -2459,6 +1577,14 @@ function hasSignificantWorkingTreeChanges(worktreePath) {
   return false;
 }
 
+function originRemoteLooksLikeGithub(repoRoot) {
+  const originUrl = readGitConfig(repoRoot, 'remote.origin.url');
+  if (!originUrl) {
+    return false;
+  }
+  return /github\.com[:/]/i.test(originUrl);
+}
+
 function autoFinishReadyAgentBranches(repoRoot, options = {}) {
   const baseBranch = String(options.baseBranch || '').trim();
   const dryRun = Boolean(options.dryRun);
@@ -2655,14 +1781,6 @@ function writeProtectedBranches(repoRoot, branches) {
   gitRun(repoRoot, ['config', GIT_PROTECTED_BRANCHES_KEY, branches.join(' ')]);
 }
 
-function readGitConfig(repoRoot, key) {
-  const result = gitRun(repoRoot, ['config', '--get', key], { allowFailure: true });
-  if (result.status !== 0) {
-    return '';
-  }
-  return (result.stdout || '').trim();
-}
-
 function resolveBaseBranch(repoRoot, explicitBase) {
   if (explicitBase) {
     return explicitBase;
@@ -2679,18 +1797,6 @@ function resolveSyncStrategy(repoRoot, explicitStrategy) {
     throw new Error(`Invalid sync strategy '${strategy}' (expected: rebase or merge)`);
   }
   return strategy;
-}
-
-function currentBranchName(repoRoot) {
-  const result = gitRun(repoRoot, ['branch', '--show-current'], { allowFailure: true });
-  if (result.status !== 0) {
-    throw new Error('Unable to detect current branch');
-  }
-  const branch = (result.stdout || '').trim();
-  if (!branch) {
-    throw new Error('Detached HEAD is not supported for sync operations');
-  }
-  return branch;
 }
 
 function repoHasHeadCommit(repoRoot) {
@@ -2753,23 +1859,6 @@ function printSetupRepoHints(repoRoot, baseBranch, repoLabel = '') {
   }
 }
 
-function workingTreeIsDirty(repoRoot) {
-  const result = gitRun(repoRoot, ['status', '--porcelain'], { allowFailure: true });
-  if (result.status !== 0) {
-    throw new Error('Unable to inspect git working tree status');
-  }
-  const lines = (result.stdout || '').split('\n').filter((line) => line.length > 0);
-  const significant = lines.filter((line) => {
-    const pathPart = (line.length > 3 ? line.slice(3) : '').trim();
-    if (!pathPart) return false;
-    if (pathPart === LOCK_FILE_RELATIVE) return false;
-    if (pathPart.startsWith(`${LOCK_FILE_RELATIVE} -> `)) return false;
-    if (pathPart.endsWith(` -> ${LOCK_FILE_RELATIVE}`)) return false;
-    return true;
-  });
-  return significant.length > 0;
-}
-
 function ensureOriginBaseRef(repoRoot, baseBranch) {
   const fetch = gitRun(repoRoot, ['fetch', 'origin', baseBranch, '--quiet'], { allowFailure: true });
   if (fetch.status !== 0) {
@@ -2783,19 +1872,6 @@ function ensureOriginBaseRef(repoRoot, baseBranch) {
   if (hasRemoteBase.status !== 0) {
     throw new Error(`Remote base branch not found: origin/${baseBranch}`);
   }
-}
-
-function aheadBehind(repoRoot, branchRef, baseRef) {
-  const result = gitRun(repoRoot, ['rev-list', '--left-right', '--count', `${branchRef}...${baseRef}`], {
-    allowFailure: true,
-  });
-  if (result.status !== 0) {
-    throw new Error(`Unable to compute ahead/behind for ${branchRef} vs ${baseRef}`);
-  }
-  const parts = (result.stdout || '').trim().split(/\s+/).filter(Boolean);
-  const ahead = Number.parseInt(parts[0] || '0', 10);
-  const behind = Number.parseInt(parts[1] || '0', 10);
-  return { ahead: Number.isFinite(ahead) ? ahead : 0, behind: Number.isFinite(behind) ? behind : 0 };
 }
 
 function lockRegistryStatus(repoRoot) {
@@ -2989,13 +2065,6 @@ function claimLocksForAutoCommit(repoRoot, worktreePath, branch) {
   }
 }
 
-function branchExists(repoRoot, branch) {
-  const result = gitRun(repoRoot, ['show-ref', '--verify', '--quiet', `refs/heads/${branch}`], {
-    allowFailure: true,
-  });
-  return result.status === 0;
-}
-
 function resolveFinishBaseBranch(repoRoot, _sourceBranch, explicitBase) {
   if (explicitBase) {
     return explicitBase;
@@ -3007,22 +2076,6 @@ function resolveFinishBaseBranch(repoRoot, _sourceBranch, explicitBase) {
   }
 
   return DEFAULT_BASE_BRANCH;
-}
-
-function branchMergedIntoBase(repoRoot, branch, baseBranch) {
-  if (!branchExists(repoRoot, baseBranch)) {
-    return false;
-  }
-  const result = gitRun(repoRoot, ['merge-base', '--is-ancestor', branch, baseBranch], {
-    allowFailure: true,
-  });
-  if (result.status === 0) {
-    return true;
-  }
-  if (result.status === 1) {
-    return false;
-  }
-  throw new Error(`Unable to determine merge status for ${branch} -> ${baseBranch}`);
 }
 
 function autoCommitWorktreeForFinish(repoRoot, worktreePath, branch, options) {
@@ -4336,7 +3389,13 @@ function doctor(rawArgs) {
 
   const blocked = protectedBaseWriteBlock(singleRepoOptions, { requireBootstrap: false });
   if (blocked) {
-    runDoctorInSandbox(singleRepoOptions, blocked);
+    doctorModule.runDoctorInSandbox(singleRepoOptions, blocked, {
+      startProtectedBaseSandbox,
+      cleanupProtectedBaseSandbox,
+      ensureOmxScaffold,
+      configureHooks,
+      autoFinishReadyAgentBranches,
+    });
     return;
   }
